se_m12301126/requirements.txt
click==7.1.2
numpy==1.21.2
pathlib==1.0.1
pycryptodome==3.10.4